Added some getter methods to the MemoryCallback class.
The MemoryCallback class is used by the SequentialBuilder to build the
state space.
It already computes the sum of rates divided by target and label, and
thus it does most of the work already.
The problem is that it does not expose such information but the done
method returns a MemoryStateSpace which uses that information to compute
the *unlabeled* rates for the CTMC.

I've now added some getters in order to be able to use this
MemoryCallback class in my AggregatedBuilder.
This information can then be used to build the LTS.