Finished code to produce the aggregated lts.
The code will probably require major refactorings.
I believe a lot of information is being re-computed many times;
we have to think about which information is essential for each part of
the code and think of a way to be able to pass that information without
having to recompute it or transform it.