Fixed most of the issues with the code.
Now the code should be computing the correct aggregated LTS.
It is much faster than before, but it still uses a lot of memory, since
it needs to compute the complete state space before starting
aggregation.

Note:

  - We should really try to support aggregate arrays even in the
aggregation code. This makes it possible to handle big initial state
spaces.
  - In doesn't really make that much senso to build an aggregated LTS
and then immediately build an IStateSpace from it. We could just compute
the data and build the IStateSpace. Also, there's probably a smarter way
to compute the rates that avoid some multiple calls...