Fixed tau's handling with respect to hiding.
Before tau actions could not be handled because their action id was
defined as -1, which triggers an error when used as an index.
Moreover the code that applied the hiding rule was simply zeroing the
rate of the named action type, effectively removing the transition
instead of renaming the transition with action type tau.

Now 0 is reserved to the tau action type and when using hiding the
hiding rule is applied and the tau transitions are introduced.